Adds the logic to check and wait for all dependencies before working a job
