+++ tests/DummyFilesystemWorker.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Webgriffe\Esb;
 
+use function Amp\delay;
 use Amp\File;
 use Amp\Promise;
 use Amp\Success;
@@ -20,9 +21,15 @@ final class DummyFilesystemWorker implements WorkerInterface
      */
     private $filename;
 
-    public function __construct(string $filename)
+    /**
+     * @var int
+     */
+    private $duration;
+
+    public function __construct(string $filename, int $duration = 0)
     {
         $this->filename = $filename;
+        $this->duration = $duration;
     }
 
     /**
@@ -43,6 +50,9 @@ final class DummyFilesystemWorker implements WorkerInterface
             $content = '';
             if (yield File\exists($this->filename)) {
                 $content = yield \Amp\File\get($this->filename);
+            }
+            if ($this->duration) {
+                yield delay($this->duration * 1000);
             }
             //The date() function does not support microseconds, whereas DateTime does.
             $now = new \DateTime('now');
